[libcalamaresui] Move ExecutionViewStep where it belongs
 - viewpages/ collects the ViewStep implementations
 - chase header moving and tidy some #includes
