+++ src/libcalamaresui/viewpages/ExecutionViewStep.h
